refactor :: domain - common이 아닌 common, domain, global
