+++ src/main/kotlin/com/stack/knowledege/common/annotation/usecase/UseCase.kt
@@ -1,4 +1,4 @@
-package com.stack.knowledege.domain.common.annotation.usecase
+package com.stack.knowledege.common.annotation.usecase
 
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
@@ -6,5 +6,5 @@ import org.springframework.transaction.annotation.Transactional
 @Retention(AnnotationRetention.RUNTIME)
 @Target(AnnotationTarget.CLASS)
 @Component
-@Transactional(readOnly = true)
-annotation class ReadOnlyUseCase
+@Transactional(rollbackFor = [Exception::class])
+annotation class UseCase
