+++ src/main/kotlin/com/stack/knowledege/domain/item/application/usecase/QueryAllItemUseCase.kt
@@ -2,7 +2,7 @@ package com.stack.knowledege.domain.item.application.usecase
 
 import com.stack.knowledege.domain.item.application.spi.QueryItemPort
 import com.stack.knowledege.domain.item.presentation.data.response.ItemResponse
-import com.stack.knowledege.domain.common.annotation.usecase.ReadOnlyUseCase
+import com.stack.knowledege.common.annotation.usecase.ReadOnlyUseCase
 
 @ReadOnlyUseCase
 class QueryAllItemUseCase(
