+++ src/main/kotlin/com/stack/knowledege/domain/mission/application/usecase/CreateMissionUseCase.kt
@@ -6,7 +6,7 @@ import com.stack.knowledege.domain.mission.domain.constant.MissionStatus
 import com.stack.knowledege.domain.mission.presentation.data.request.CreateMissionRequest
 import com.stack.knowledege.domain.user.application.spi.UserPort
 import com.stack.knowledege.domain.user.application.validator.UserValidator
-import com.stack.knowledege.domain.common.annotation.usecase.UseCase
+import com.stack.knowledege.common.annotation.usecase.UseCase
 import java.time.LocalTime
 import java.util.*
