+++ src/main/kotlin/com/stack/knowledege/domain/order/application/usecase/OrderItemUseCase.kt
@@ -10,7 +10,7 @@ import com.stack.knowledege.domain.order.presentation.data.request.OrderItemRequ
 import com.stack.knowledege.domain.student.application.spi.QueryStudentPort
 import com.stack.knowledege.domain.student.exception.StudentNotFoundException
 import com.stack.knowledege.domain.user.application.spi.QueryUserPort
-import com.stack.knowledege.domain.common.annotation.usecase.UseCase
+import com.stack.knowledege.common.annotation.usecase.UseCase
 import java.util.UUID
 
 @UseCase
