+++ src/main/kotlin/com/stack/knowledege/domain/solve/application/usecase/SolveMissionUseCase.kt
@@ -13,7 +13,7 @@ import com.stack.knowledege.domain.student.application.spi.QueryStudentPort
 import com.stack.knowledege.domain.student.exception.StudentNotFoundException
 import com.stack.knowledege.domain.user.application.spi.QueryUserPort
 import com.stack.knowledege.domain.user.domain.constant.Authority
-import com.stack.knowledege.domain.common.annotation.usecase.UseCase
+import com.stack.knowledege.common.annotation.usecase.UseCase
 import java.util.UUID
 
 @UseCase
