+++ src/main/kotlin/com/stack/knowledege/domain/student/persistence/entity/StudentJpaEntity.kt
@@ -1,7 +1,7 @@
 package com.stack.knowledege.domain.student.persistence.entity
 
 import com.stack.knowledege.domain.user.persistence.entity.UserJpaEntity
-import com.stack.knowledege.domain.common.entity.BaseIdEntity
+import com.stack.knowledege.common.entity.BaseIdEntity
 import java.util.*
 import javax.persistence.Column
 import javax.persistence.Entity
