+++ src/main/kotlin/com/stack/knowledege/domain/user/application/usecase/QueryScoringPageUseCase.kt
@@ -2,7 +2,7 @@ package com.stack.knowledege.domain.user.application.usecase
 
 import com.stack.knowledege.domain.solve.application.spi.QuerySolvePort
 import com.stack.knowledege.domain.user.presentation.data.response.AllScoringResponse
-import com.stack.knowledege.domain.common.annotation.usecase.ReadOnlyUseCase
+import com.stack.knowledege.common.annotation.usecase.ReadOnlyUseCase
 
 @ReadOnlyUseCase
 class QueryScoringPageUseCase(
