+++ src/main/kotlin/com/stack/knowledege/domain/user/domain/User.kt
@@ -1,7 +1,7 @@
 package com.stack.knowledege.domain.user.domain
 
 import com.stack.knowledege.domain.user.domain.constant.Authority
-import com.stack.knowledege.domain.common.annotation.Aggregate
+import com.stack.knowledege.common.annotation.Aggregate
 import java.util.*
 
 @Aggregate
